requirements.txt
allure-pytest
allure-python-commons
pycodestyle
pytest
pytest-xdist
requests
selenium
faker